feat(rpc): separate rpc call for transition to 2.0 rpc call
