+++ src/core/atomicdex/api/mm2/rpc.withdraw.cpp
@@ -1,0 +1,58 @@
+//
+// Created by Sztergbaum Roman on 08/06/2021.
+//
+
+//! Deps
+#include <nlohmann/json.hpp>
+
+//! Our Headers
+#include "rpc.withdraw.hpp"
+
+namespace mm2::api
+{
+    void
+    to_json(nlohmann::json& j, const withdraw_fees& cfg)
+    {
+        j["type"] = cfg.type;
+        if (cfg.type == "EthGas")
+        {
+            j["gas"]       = cfg.gas_limit.value_or(55000);
+            j["gas_price"] = cfg.gas_price.value();
+        }
+        else if (cfg.type == "Qrc20Gas")
+        {
+            j["gas_limit"] = cfg.gas_limit.value_or(40);
+            j["gas_price"] = std::stoi(cfg.gas_price.value());
+        }
+        else
+        {
+            j["amount"] = cfg.amount.value();
+        }
+    }
+
+    void
+    to_json(nlohmann::json& j, const withdraw_request& cfg)
+    {
+        j["coin"]   = cfg.coin;
+        j["amount"] = cfg.amount;
+        j["to"]     = cfg.to;
+        j["max"]    = cfg.max;
+        if (cfg.fees.has_value())
+        {
+            j["fee"] = cfg.fees.value();
+        }
+    }
+
+    void
+    from_json(const nlohmann::json& j, withdraw_answer& answer)
+    {
+        if (j.count("error") >= 1)
+        {
+            answer.error = j.at("error").get<std::string>();
+        }
+        else
+        {
+            answer.result = j.get<transaction_data>();
+        }
+    }
+}
